moved tests from their own file to trade_executor_tests
